Refactor public API.
