+++ scripts/__init__.py
@@ -1,2 +1,19 @@
-from .wiki import *
-from .schemas import *
+from .wiki import (
+    schemas,
+    load_entities,
+    establish_db_connection,
+    extract_demo_dump,
+    load_alias_entity_prior_probabilities,
+    parse,
+    namespaces,
+)
+
+__all__ = [
+    "schemas",
+    "load_entities",
+    "establish_db_connection",
+    "extract_demo_dump",
+    "load_alias_entity_prior_probabilities",
+    "parse",
+    "namespaces",
+]
